+++ app/main/controller/video_analysis.py
@@ -1,0 +1,3 @@
+from flask import Blueprint, request, json, Response
+
+api = Blueprint("video_analysis", __name__, url_prefix="/video_analysis")
